% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMU-package.R
\docType{data}
\name{all_catchments}
\alias{all_catchments}
\title{All catchments}
\format{
A spatial feature containing 130 features
\describe{
\item{fid}{Feature ID of the polygon. Primary key differentiating each unique polygon and used to match radar survey data to catchments.}
\item{region}{MAMU conservation region the catchment falls within.}
\item{code}{Short alphanumeric code abbreviating the catchment name.}
\item{catchment}{Name of the catchment.}
\item{version}{Version number of the catchment. Useful to keep track of cases where the catchment has multiple iterations and been re-drawn (e.g., see Artlish or Toquart for an example).}
\item{coastal_buff_500m}{Logical (T/F) - does the catchment have a 500m buffer from the coastline?}
\item{area}{Polygon area in m^2.}
}
}
\usage{
all_catchments
}
\description{
A spatial feature (\code{sf}) class object containing a multipolygon
of all versions of all MAMU catchments as of 2023-03-31.
}
\keyword{data}
\keyword{feature,}
\keyword{spatial}
