% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{BC_DEM}
\alias{BC_DEM}
\title{Download British Columbia Digital Elevation Model data}
\usage{
BC_DEM(
  letterblock,
  save_output = TRUE,
  overwrite = FALSE,
  output_dir = NA,
  filename = NA
)
}
\arguments{
\item{letterblock}{Character vector of map tiles of DEM data to download, e.g. \code{'92b'} for a single tile or \code{c('92b', '92g')} for multiple tiles. See the BC Maps & Orthos Base Map Online Store here for a visual: \url{https://a100.gov.bc.ca/ext/mtec/public/products/mapsheet}}

\item{save_output}{Logical (T/F). Should the downloaded DEM be saved to your local disk? Default TRUE.}

\item{overwrite}{Logical (T/F). Should the downloaded DEM overwrite any existing BC DEM data on the disk, if already present? Default FALSE.}

\item{output_dir}{Output directory to save downloaded DEM to. Defaults to current working directory.}

\item{filename}{Filename of saved DEM. By default, the filename is either 1) the letterblock names, seperated by underscores (e.g., "92b_92g.dem"), if there are fewer than 5 letterblocks were downloaded; or 2) simply "BC_DEM.dem", if greater than 5 letterblocks were downloaded.}
}
\value{
A RasterSpat object.
}
\description{
Download British Columbia Digital Elevation Model data
}
\examples{
\dontrun {
BC_DEM("92b", save_output = F) # download Victoria area DEM
BC_DEM(c("92b", "92g", "92f", "92c"), output_dir = "temp/bc_dem") # download SW Vancouer island + lower mainland DEM and save output to the "temp/bc_dem" directory. The generated file will be "temp/bc_dem/92b_92c_92f_92g.dem"
BC_DEM(c("103k", "103j", "103f", "103g", "103c", "103b", "102o"), output_dir = "temp/bc_dem", filename = "haida_gwaii.dem") # download Haida Gwaii DEM and save output to "temp/bc_dem/haida_gwaii.dem"
}
}
