% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMU-package.R
\docType{data}
\name{catch_combos}
\alias{catch_combos}
\title{Radar site survey - Catchment combinations}
\format{
A dataframe with 226 rows and 2 variables:
\describe{
\item{new_name}{New, cleaned site name as generated by \code{\link[=process_radar_data]{process_radar_data()}}. This cleaned name is associated with a shapefile feature id (\code{fid}) in the \code{all_catchments} geopackage.}
\item{fid}{Feature ID of the associated catchment polygon feature within \link{all_catchments}.}
}
}
\usage{
catch_combos
}
\description{
A small lookup table containing all survey location names and the
feature id (\code{fid}) of the associated catchment polygon feature in the \code{all_catchments.gpkg}
geopackage file. This table is used to match survey data to the correct catchment.
}
\keyword{data}
