% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_zip}
\alias{download_zip}
\title{Download zipped DEM file from a given URL}
\usage{
download_zip(url, dir)
}
\arguments{
\item{url}{Download link to a zipped DEM file (typically as returned by \link{tile_urls})}

\item{dir}{Output directory to save the downloaded DEM ZIP file to}
}
\value{

}
\description{
Download zipped DEM file from a given URL
}
