% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{make_vrt}
\alias{make_vrt}
\title{Virtual Raster Dataset}
\usage{
make_vrt(path, filename, overwrite = FALSE)
}
\arguments{
\item{path}{Path to directory containing DEM files you wish to assemble into a VRT}

\item{filename}{Output VRT filename}

\item{overwrite}{Logical (T/F). Should \code{filename} be overwritten if it exists?}
}
\value{
A SpatRaster object.
}
\description{
Create VRT from a list of DEM files. This function is a simple wrapper of the \code{\link[terra:vrt]{terra::vrt()}} function.
}
\examples{
\dontrun{
# Download some BC DEM tiles...
tiles <- c("92h", "93c", "93l", "93m", "104a", "104b")
sapply(tiles, BC_DEM,
       save_output = TRUE,
       overwrite = FALSE,
       output_dir = "temp/DEM_tiles/")
# Create VRT
make_vrt(path = "temp/DEM_tiles/", filename = "temp/BC_DEM_VRT.vrt", overwrite = TRUE)
}
}
