% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMU-package.R
\docType{package}
\name{MAMU-package}
\alias{MAMU}
\alias{MAMU-package}
\title{MAMU: Functions for BC Marbled Murrelet (MAMU) Forestry analyses}
\description{
This package includes functions to read and clean MAMU datasets (e.g., radar surveys or VRI forest cover) and then perform analyses (e.g., population-to-habitat area trends).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/popovs/MAMU}
  \item Report bugs at \url{https://github.com/popovs/MAMU/issues}
}

}
\author{
\strong{Maintainer}: Sarah Popov \email{sarah.popov@mail.mcgill.ca} (\href{https://orcid.org/0000-0002-8257-6438}{ORCID})

}
\keyword{internal}
