% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.R
\name{process_radar_data}
\alias{process_radar_data}
\title{Import and Clean ECCC-FLNR MAMU Radar Data}
\usage{
process_radar_data(path, site_groupings)
}
\arguments{
\item{path}{File path to radar data Excel file.}

\item{site_groupings}{(optional) Dataframe of site name groups with two columns: 1) \code{original_name} and 2) \code{new_name}.}
}
\value{
A dataframe of cleaned radar data.
}
\description{
This function will read in the Excel file from the provided path and
perform a standardized series of cleaning steps that always need to be performed
on ECCC-FLNR MAMU radar survey data prior to any analyses.
}
\details{
The function first performs a few housekeeping tasks:
\itemize{
\item Make R-friendly column names
\item Create the \code{survey_date} column from the year, month, and day columns
\item Re-calculate day-of-year (\code{doy})
\item Fix minor typos, e.g. "Skwakwa" -> "Skwawka"
}

It then filters down the data with the following criteria:
\itemize{
\item Include only "Complete" surveys
\item Include only pre-dawn surveys (i.e., after 2 am and before 8 am)
\item Exclude August surveys
\item Exclude Alaska border region surveys
}

Finally, it creates the \code{new_name} column, which groups sites with slightly
different names that should really be a single site. The user can either rely
on the default radar site name changes that are bundled in this package (see \code{?rs}),
or they can specify their own as a dataframe in the \code{site_groupings} param.
}
\section{Site groupings}{
Site groupings must be provided as a dataframe with two columns: \code{original_name}
and \code{new_name}, where the \code{original_name} column contains site names found in the
raw data and the \code{new_name} column contains name you want the site to be renamed to.
Run \code{data(rs)} for an example of the default name changes bundled with this package.
}

\examples{
# Clean using function defaults
process_radar_data("~/Documents/path/to/radar/data.xlsx")

# Provide your own site groupings
sg <- data.frame(original_name = c("Fairfax 2", "Coleman Boat", "Nekite 1", "Nekite 2"),
                 new_name = c("Fairfax", "Coleman", "Nekite", "Nekite"))
process_radar_data("~/Documents/path/to/radar/data.xlsx",
                   site_groupings = sg)

}
