% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{radar_buffers}
\alias{radar_buffers}
\title{Buffers around radar survey sites}
\usage{
radar_buffers(surveys)
}
\arguments{
\item{surveys}{A dataframe of cleaned radar survey data; i.e., the output from \code{\link[=process_radar_data]{process_radar_data()}}.}
}
\value{
An sf object
}
\description{
Calculate 10, 20, 30, 40, and 50 km buffers around a given set of
radar survey sites.
}
\examples{
surveys <- process_radar_data("~/Documents/path/to/radar/data.xlsx")
buffs <- radar_buffers(surveys)
plot(sf::st_geometry(buffs))
}
