% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_cone.R
\name{radar_cone}
\alias{radar_cone}
\title{Create a conical gradient of interpolated values around a point}
\usage{
radar_cone(
  pt,
  radius,
  heading,
  theta = 0,
  res = 500,
  mask = TRUE,
  invert = FALSE,
  maxvalue = 1
)
}
\arguments{
\item{pt}{A single (X,Y) coordinate of class \code{sfc_POINT} (created by the \code{sf} package), where the CRS is a projected coordinate system.}

\item{radius}{Numeric. Distance in meters from the point \code{pt} to draw the circle radius.}

\item{heading}{Numeric. Heading, in compass degrees, which to interpolate values from.}

\item{theta}{Numeric (0 to 360). Angle, in degrees, indicating the amount of the circle that should be taken up by the \code{maxvalue}. For example, if you want to create a wedge where 1/3 of the circle == 1, set \code{theta = 120}. Default \code{0}.}

\item{res}{Numeric. Resolution, in meters, of the output raster. Default \code{500}.}

\item{mask}{Boolean (TRUE/FALSE). Should the output be masked to a circle shape (rather than square)? Default \code{TRUE}.}

\item{invert}{Boolean (TRUE/FALSE). Should the values of the output cone be inverted, such that the values in the direction of the heading are 0 rather than 1? Default \code{FALSE}.}

\item{maxvalue}{Numeric. Maximum value of the interpolated values of the cone. The interpolated values run from 0 to \code{maxvalue}. Default \code{1}.}
}
\value{
A \code{SpatRaster} object
}
\description{
Using a given coordinate, this function will generate a circle
of a given radius around the point and fill the circle with a
gradient of values in a conical pattern. By default, the interpolated
values range from zero to one, with values of one point in the direction
of the specified heading.
}
\details{
When using this function with spatial features,
ensure your data are projected in an equal-area coordinate system. This
function assumes meters as the default unit.

This function was originally developed with the intention of
generating 'cones of probability' for estimating the dispersal
of birds flying across a landscape. Radar stations based across British
Columbia, Canada, collected data on the mean headings of birds flying
across a radar station. These headings indicate the more likely direction
birds will be flying to, while headings in the opposite direction
indicate areas the birds are flying away from. Thus, the gradient of
values from 0 (flying away) to 1 (flying towards) was developed around
each given radar station using this function.
}
\examples{
library(magrittr)
pt <- sf::st_point(c(-126, 53)) \%>\%
  sf::st_sfc(crs = 4326) \%>\%
  sf::st_transform(crs = 3005)

radius <- 30000 # 30 km
heading <- 45 # NE direction

terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = heading))

# Adjust heading
terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = 120))

# Check values ≈ 1
terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = 120) > 0.99)

# Increase wedge area where values ≈ 1 to 1/4 of the circle
terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = 120,
                       theta = 90) > 0.99)

# Increase wedge area where values ≈ 1 to 1/3 of the circle
terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = 120,
                       theta = 120) > 0.99)

Invert the cone
terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = 120,
                       theta = 90,
                       invert = TRUE))

terra::plot(radar_cone(pt = pt,
                       radius = radius,
                       heading = 120,
                       theta = 120,
                       invert = TRUE))
}
