% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rasterize_vri}
\alias{rasterize_vri}
\title{Rasterize VRI polygons}
\usage{
rasterize_vri(path, layer, below_900m = TRUE, res = 10)
}
\arguments{
\item{path}{Path to either .gdb directory, .shp file, or .gpkg containing the VRI polygons.}

\item{layer}{If path is a Geodatabase or Geopackage, layer name of the VRI polygons.}

\item{below_900m}{Logical (T/F). Should the function only rasterize polygons below 900m? Default \code{TRUE}. If \code{FALSE}, the function will ignore the 'Elevation' column and rasterize all polygons.}

\item{res}{Resolution in meters of the resultant raster. Default 10 m^2.}
}
\value{

}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[=rasterize_vri]{rasterize_vri()}} will quickly rasterize a given VRI (multi)polygon using \code{\link[fasterize:fasterize]{fasterize::fasterize()}}
under the hood at a default resolution of 10x10 m per pixel. By default, the function
filters out any polygons occurring >900m elevation as reported in the VRI 'Elevation' column.
}
\examples{
path <- "~/Documents/path/to/VRI.gdb"
layer <- "VRI_Algorithm_Jan11_2023"
vri <- rasterize_vri(path = path, layer = layer)
}
