% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMU-package.R
\docType{data}
\name{rs}
\alias{rs}
\title{Radar site groupings}
\format{
A dataframe with 15 rows and 2 variables:
\describe{
\item{original_name}{Original site name as appears in the 'Name$' column in the raw ECCC-FLNR MAMU radar data.}
\item{new_name}{New, cleaned site name. This cleaned name can then be used in later analyses to group radar sites that need to be consolidated.}
}
}
\usage{
rs
}
\description{
A small data lookup containing radar survey sites that should be renamed
during the data cleaning process in \code{\link[=process_radar_data]{process_radar_data()}}.
}
\keyword{data}
