% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tile_urls}
\alias{tile_urls}
\title{Return letterblock (tile) download URLs}
\usage{
tile_urls(url = "https://pub.data.gov.bc.ca/datasets/175624/", tile)
}
\arguments{
\item{url}{FTP URL containing all parent DEM download directories (i.e., https://pub.data.gov.bc.ca/datasets/175624/)}

\item{tile}{Name of letterblock that you want to download DEMs from (e.g., '95d', '103p', '115a', etc.)}
}
\value{
A character vector of all download URLs associated with a given BC DEM letterblock.
}
\description{
For a given letterblock (aka tile), scrape all the associated download URLs.
}
