% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{vri_dir}
\alias{vri_dir}
\title{Output path to the VRI database}
\usage{
vri_dir(...)
}
\arguments{
\item{...}{arguments potentially passed to \code{rappdirs::user_data_dir}.}
}
\description{
Prints the VRI database storage location, regardless of OS.
E.g., for Mac:

\verb{~Library/Application Support/VRI}

If on Windows, either:

\verb{C:\\\\Documents and Settings\\<User>\\\\Application Data\\\\Local Settings\\<AppAuthor>\\\\VRI}

Or:

\verb{C:\\\\Documents and Settings\\<User>\\\\Application Data\\<AppAuthor>\\\\VRI}
}
\examples{
\dontrun{
bpfa_dir()
}
}
